front: refacto create consts file + remove regex for util function
